[map] clear the gps tracker on stop
The prev solution was to clean the tracker points before starting a new track recording.
This cause the bug when if we starts the recording the `TrackRecordingUpdateHandler` is called by subscribers but receives the track info from the previous track. It happens because the starting is an async method and it clean up the gps collection a little bit later. There are no reason to store all the recorded data in memory until the new recording begins.

Signed-off-by: Kiryl Kaveryn <kirylkaveryn@gmail.com>